+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Dropdown's selected option when value is `undefined`.
+
 ## [9.112.27] - 2020-03-26
 
 ### Fixed